Optimized rb-find by taking advantage of the sentinel.
When using a leaf sentinel, you can set the key of the sentinel to the
key you are looking for. Then you no longer have to explicitly check
if you are looking at a leaf while traversing the tree. You only have
to check when you find a node with the same key.